Centralize card-schedule snapshot types and replay logic
Moves CardScheduleSnapshot and RemovedFavoriteSnapshot into a dedicated
file and exposes a single FavoritesQueries.applyCardScheduleSnapshots
extension that both restoreForUndo and restorePausedWord call. Drops
PausedWordSnapshot — it was an empty wrapper around the card list.
