+++ composeApp/src/commonMain/kotlin/com/slovy/slovymovyapp/data/learning/session/SessionService.kt
@@ -6,6 +6,7 @@ import com.slovy.slovymovyapp.analytics.use
 import com.slovy.slovymovyapp.analytics.useWithResult
 import com.slovy.slovymovyapp.data.Language
 import com.slovy.slovymovyapp.data.favorites.CardScheduleSnapshot
+import com.slovy.slovymovyapp.data.favorites.applyCardScheduleSnapshots
 import com.slovy.slovymovyapp.data.learning.*
 import com.slovy.slovymovyapp.data.learning.fsrs.CrossFamilyCredit
 import com.slovy.slovymovyapp.data.learning.fsrs.FsrsConfig
@@ -28,8 +29,6 @@ import kotlin.time.*
 import kotlin.time.Duration.Companion.days
 import kotlin.time.Duration.Companion.hours
 import kotlin.uuid.Uuid
-
-data class PausedWordSnapshot(val cards: List<CardScheduleSnapshot>)
 
 @OptIn(ExperimentalTime::class)
 class SessionService(
@@ -204,7 +203,7 @@ class SessionService(
         setCardAvailableAfter(card.card)
     }
 
-    suspend fun suspendWord(card: SessionCard, duration: Duration): PausedWordSnapshot =
+    suspend fun suspendWord(card: SessionCard, duration: Duration): List<CardScheduleSnapshot> =
         withContext(Dispatchers.IO) {
             val availableAfter = clock.now().toEpochMilliseconds() + duration.inWholeMilliseconds
             learning.transactionWithResult {
@@ -223,19 +222,13 @@ class SessionService(
                     lang_code = card.card.langCode,
                     lemma_id = card.card.lemmaId,
                 )
-                PausedWordSnapshot(cards = cards)
+                cards
             }
         }
 
-    suspend fun restorePausedWord(snapshot: PausedWordSnapshot) = withContext(Dispatchers.IO) {
+    suspend fun restorePausedWord(snapshot: List<CardScheduleSnapshot>) = withContext(Dispatchers.IO) {
         learning.transaction {
-            for (card in snapshot.cards) {
-                learning.restoreCardScheduling(
-                    suspended = card.suspended,
-                    available_after = card.availableAfter,
-                    id = card.id,
-                )
-            }
+            learning.applyCardScheduleSnapshots(snapshot)
         }
     }
